Refactor: add UnsafeSerialization.writeIntsWithPrefix method to simplify a couple LSH models (#593)
