+++ docs/pages/performance/fashion-mnist/results.md
@@ -1,10 +1,10 @@
 |Model|Parameters|Recall|Queries per Second|
 |---|---|---|---|
-|eknn-l2lsh|L=100 k=4 w=1024 candidates=500 probes=0|0.378|351.096|
-|eknn-l2lsh|L=100 k=4 w=1024 candidates=1000 probes=0|0.447|291.666|
-|eknn-l2lsh|L=100 k=4 w=1024 candidates=500 probes=3|0.634|277.702|
-|eknn-l2lsh|L=100 k=4 w=1024 candidates=1000 probes=3|0.716|238.914|
-|eknn-l2lsh|L=100 k=4 w=2048 candidates=500 probes=0|0.767|288.441|
-|eknn-l2lsh|L=100 k=4 w=2048 candidates=1000 probes=0|0.847|246.201|
-|eknn-l2lsh|L=100 k=4 w=2048 candidates=500 probes=3|0.922|192.499|
-|eknn-l2lsh|L=100 k=4 w=2048 candidates=1000 probes=3|0.960|177.009|
+|eknn-l2lsh|L=100 k=4 w=1024 candidates=500 probes=0|0.378|363.121|
+|eknn-l2lsh|L=100 k=4 w=1024 candidates=1000 probes=0|0.446|299.144|
+|eknn-l2lsh|L=100 k=4 w=1024 candidates=500 probes=3|0.634|270.522|
+|eknn-l2lsh|L=100 k=4 w=1024 candidates=1000 probes=3|0.716|240.419|
+|eknn-l2lsh|L=100 k=4 w=2048 candidates=500 probes=0|0.768|280.053|
+|eknn-l2lsh|L=100 k=4 w=2048 candidates=1000 probes=0|0.847|240.014|
+|eknn-l2lsh|L=100 k=4 w=2048 candidates=500 probes=3|0.922|186.668|
+|eknn-l2lsh|L=100 k=4 w=2048 candidates=1000 probes=3|0.960|166.241|
